Updated App.js and removed Home
